new test with folding at the poles
